requirements.txt
pyshark
scapy
numpy
requests
geoip2
dnspython
cryptography
